Added requirements.txt and .gitignore files
Created a requirements.txt file to specify the dependencies required for running the project.

Added pygame==2.5.2 and numpy as dependencies to ensure that the project runs smoothly.

Included an empty .gitignore file to specify files and directories that should be ignored by Git.

An empty .gitignore file allows for customization in the future to exclude specific files or directories from version control.