Added 'levels' folder and 'level_1.py' file
Created a folder named 'levels' to organize game level-related files.

Added a Python file named 'level_1.py' within the 'levels' folder to define properties for the first level of the game.
In 'level_1.py':

Defined a class named 'maps' to encapsulate properties of the game map.

Defined lists for 'pos', 'width', 'height', and 'updown' to store positions, dimensions, and movement direction of obstacles in the level.
'pos': Represents the x-coordinate positions of obstacles.
'width': Represents the width of obstacles.
'height': Represents the height of obstacles.
'updown': Represents the movement direction of obstacles (1 for upward, 0 for downward).